Accept int keys for `features` dict in ServingInputReceiver
In particular, proto enum values are represented as ints in Python and can be useful as safer keys that are recognized and linked by IDEs.

The `labels` dict is treated like the `features` dict. However, this does *not* change the allowed keys for the receiver_tensor dicts, which are part of the model's external-facing signature and should remain string keys (descriptive names).

PiperOrigin-RevId: 246993535